+++ src/core/scope/index.ts
@@ -1,1 +1,2 @@
-export {};
+export * from "./claim-scope.js";
+export * from "./scope-types.js";